feat: 카테고리맵핑 조회 API API 추가 및 S3 코드 정리

+++ database/migrations/2024_03_06_185220_create_category_mappings_table.php
@@ -16,7 +16,7 @@ class CreateCategoryMappingsTable extends Migration
     {
         Schema::create('category_mappings', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('category_id')->nullable(false)->comment('카테고리ID');
+            $table->unsignedBigInteger('category_id')->nullable(false)->comment('카테고리ID');
             $table->string('mapping_channel', 50)->nullable(false)->default("onchannel")->comment('카테고리 맵핑 채널');
             $table->string('mapping_code', 50)->nullable(false)->comment('카테고리 맵핑 코드');
